Moving spec files to specific folders
