+++ test/spec/quarters/DeleteQuarter.controller.spec.ts
@@ -1,9 +1,9 @@
 import { app } from 'infra/http/app';
 import supertest from 'supertest';
-import { deleteAll, disconnect } from '../../../../test/database';
-import { createAcademicYear } from '../../../../test/entities/AcademicYearFactory';
-import { createQuarter } from '../../../../test/entities/QuarterFactory';
-import { createUser, authenticateUser } from '../../../../test/entities/UserFactory';
+import { deleteAll, closeConnection } from '../../database';
+import { createAcademicYear } from '../../entities/AcademicYearFactory';
+import { createQuarter } from '../../entities/QuarterFactory';
+import { createUser, authenticateUser } from '../../entities/UserFactory';
 
 describe('Delete Quarter (e2e)', () => {
   beforeAll(async () => {
@@ -11,7 +11,7 @@ describe('Delete Quarter (e2e)', () => {
   });
 
   afterAll(async () => {
-    await disconnect();
+    await closeConnection();
   });
 
   it('should delete a quarter successfully', async () => {
