+++ test/spec/tasks/GetTasksBySubjectId.controller.spec.ts
@@ -1,11 +1,11 @@
 import { app } from 'infra/http/app';
 import supertest from 'supertest';
-import { deleteAll, disconnect } from '../../../../test/database';
-import { createAcademicYear } from '../../../../test/entities/AcademicYearFactory';
-import { createQuarter } from '../../../../test/entities/QuarterFactory';
-import { createSubject } from '../../../../test/entities/SubjectFactory';
-import { createTask } from '../../../../test/entities/TaskFactory';
-import { createUser, authenticateUser } from '../../../../test/entities/UserFactory';
+import { deleteAll, closeConnection } from '../../database';
+import { createAcademicYear } from '../../entities/AcademicYearFactory';
+import { createQuarter } from '../../entities/QuarterFactory';
+import { createSubject } from '../../entities/SubjectFactory';
+import { createTask } from '../../entities/TaskFactory';
+import { createUser, authenticateUser } from '../../entities/UserFactory';
 
 describe('Get tasks by subject Id (e2e)', () => {
   beforeAll(async () => {
@@ -13,7 +13,7 @@ describe('Get tasks by subject Id (e2e)', () => {
   });
 
   afterAll(async () => {
-    await disconnect();
+    await closeConnection();
   });
 
   it('Should return a task successfully using the subject id', async () => {
